refactor(Paths): remove BasePath interface usage and handle relative compare paths
