--- src/Ml.Cli.WebApp/BasePath/IBasePath.cs
@@ -1,7 +1,0 @@
-﻿namespace Ml.Cli.WebApp.BasePath
-{
-    public interface IBasePath
-    {
-        bool IsPathSecure(string path);
-    }
-}
