+++ src/Ml.Cli.WebApp/Controllers/FilesController.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.StaticFiles;
 using Ml.Cli.FileLoader;
-using Ml.Cli.WebApp.BasePath;
+using Ml.Cli.WebApp.Paths;
 
 namespace Ml.Cli.WebApp.Controllers
 {
@@ -13,10 +13,10 @@ namespace Ml.Cli.WebApp.Controllers
     public class FilesController : ControllerBase
     {
         private readonly IFileLoader _fileLoader;
-        private readonly IBasePath _basePath;
-        private readonly ComparesPaths.ComparesPaths _comparesPaths;
+        private readonly BasePath _basePath;
+        private readonly ComparesPaths _comparesPaths;
 
-        public FilesController(IFileLoader fileLoader, IBasePath basePath, ComparesPaths.ComparesPaths comparesPaths)
+        public FilesController(IFileLoader fileLoader, BasePath basePath, ComparesPaths comparesPaths)
         {
             _fileLoader = fileLoader;
             _basePath = basePath;
