+++ src/Ml.Cli.WebApp/Program.cs
@@ -5,8 +5,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.CommandLineUtils;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Ml.Cli.WebApp.BasePath;
-using Ml.Cli.WebApp.ComparesPaths;
+using Ml.Cli.WebApp.Paths;
 
 namespace Ml.Cli.WebApp
 {
@@ -88,8 +87,8 @@ namespace Ml.Cli.WebApp
                 {
                     services.AddSingleton<IHostedService>(provider =>
                         new Worker(args.Skip(4).ToArray()));
-                    services.AddSingleton<IBasePath, BasePath.BasePath>(provider => new BasePath.BasePath(args[1]));
-                    services.AddSingleton(provider => new ComparesPaths.ComparesPaths(args[3]));
+                    services.AddSingleton(provider => new BasePath(args[1]));
+                    services.AddSingleton(provider => new ComparesPaths(args[3]));
                 });
     }
 }
