+++ tests/Ml.Cli.WebApp.Tests/ComparesControllerUnitTest.cs
@@ -2,8 +2,8 @@
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Ml.Cli.FileLoader;
-using Ml.Cli.WebApp.BasePath;
 using Ml.Cli.WebApp.Controllers;
+using Ml.Cli.WebApp.Paths;
 using Moq;
 using Newtonsoft.Json;
 using Xunit;
@@ -30,10 +30,10 @@ namespace Ml.Cli.WebApp.Tests
             fileLoader.Setup(foo => foo.ReadAllTextInFileAsync("someFolder\\\\compareFile.json")).ReturnsAsync(jsonCompareContent);
             fileLoader.Setup(foo => foo.WriteAllTextInFileAsync("someFolder\\\\compareFile.json", It.IsAny<string>()));
 
-            var basePath = new Mock<IBasePath>();
+            var basePath = new Mock<BasePath>();
             basePath.Setup(foo => foo.IsPathSecure(It.IsAny<string>())).Returns(true);
             
-            var comparesPath = new ComparesPaths.ComparesPaths("");
+            var comparesPath = new ComparesPaths("");
 
             var compareController = new ComparesController(fileLoader.Object, basePath.Object, comparesPath);
             await compareController.SaveCompare(editorContent);
@@ -49,14 +49,14 @@ namespace Ml.Cli.WebApp.Tests
         [Fact]
         public async Task ShouldGetJsonFiles()
         {
-            var basePath = new Mock<IBasePath>();
+            var basePath = new Mock<BasePath>();
             basePath.Setup(foo => foo.IsPathSecure(It.IsAny<string>())).Returns(true);
             var fileLoader = new Mock<IFileLoader>();
             fileLoader.Setup(foo => foo.EnumerateFiles("someFolder\\\\comparesFolder1"))
                 .Returns(new[] {"file1.json", "file2.png", "file3.json"});
             fileLoader.Setup(foo => foo.EnumerateFiles("someFolder\\\\comparesFolder2"))
                 .Returns(new[] {"file4.png", "file5.pdf"});
-            var comparesPath = new ComparesPaths.ComparesPaths("someFolder\\\\comparesFolder1,someFolder\\\\comparesFolder2");
+            var comparesPath = new ComparesPaths("someFolder\\\\comparesFolder1,someFolder\\\\comparesFolder2");
             
             var compareController = new ComparesController(fileLoader.Object, basePath.Object, comparesPath);
